backend/requirements.txt
boto3==1.34.34
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
